Update script to avoid deleting alembic revision

+++ app/api/orm/sql_scripts/refresh_db.sql
@@ -8,6 +8,8 @@ DECLARE
 BEGIN
     FOR v_table_name IN (SELECT table_name FROM information_schema.tables WHERE table_schema = 'public' AND table_type = 'BASE TABLE')
     LOOP
-        EXECUTE 'DELETE FROM ' || v_table_name || ' CASCADE;';
+        IF v_table_name <> 'alembic_version' THEN
+            EXECUTE 'DELETE FROM ' || v_table_name || ' CASCADE;';
+        END IF;
     END LOOP;
 END $$;
